client/src/app/services/configuration/configuration.service.ts
import { Injectable } from '@angular/core';
import { GameSelectionService } from '@app/services/game-selection/game-selection.service';
import { SocketService } from '@app/services/socket/socket.service';
import { ConfigurationEvent } from '@common/configuration.events';

@Injectable({
    providedIn: 'root',
})
export class ConfigurationService {
    constructor(private socketService: SocketService, private gameSelectionService: GameSelectionService) {
        this.socketService.on(ConfigurationEvent.ReinitializeWasDone, () => {
            this.gameSelectionService.fetchGames();
        });
    }

    reinitializeScores(gameId: string) {
        this.socketService.send(ConfigurationEvent.ReinitializeScores, { gameId });
    }

    reinitializeAllScores() {
        this.socketService.send(ConfigurationEvent.ReinitializeAllScores);
    }
}
